% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annem_utils.R
\name{annem_summary}
\alias{annem_summary}
\title{ANNEM Summary}
\usage{
annem_summary(analysis_results)
}
\arguments{
\item{analysis_results}{Results from run_annem_analysis()}
}
\description{
Print summary information about ANNEM analysis results.
}
\examples{
\dontrun{
results <- run_annem_analysis(n_agents = 100, n_steps = 50)
annem_summary(results)
}

}
