% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annem_visualization.R
\name{create_interactive_network}
\alias{create_interactive_network}
\title{Create Interactive Network Visualization}
\usage{
create_interactive_network(network, agent_types, width = 800, height = 600)
}
\arguments{
\item{network}{igraph network object}

\item{agent_types}{Character vector of agent types}

\item{width}{Width of the visualization (default: 800)}

\item{height}{Height of the visualization (default: 600)}
}
\value{
networkD3 forceNetwork object
}
\description{
Generate an interactive network visualization using networkD3.
}
\examples{
\dontrun{
library(igraph)
net <- sample_smallworld(1, 50, 4, 0.1)
agent_types <- sample(c("type1", "type2", "type3"), 50, replace = TRUE)
interactive_net <- create_interactive_network(net, agent_types)
}

}
