% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annem_utils.R
\name{fit_var_model}
\alias{fit_var_model}
\title{Fit VAR Model}
\usage{
fit_var_model(returns_data, lag_order = 1)
}
\arguments{
\item{returns_data}{Matrix of return series}

\item{lag_order}{Order of VAR model (default: 1)}
}
\value{
Fitted values from VAR model
}
\description{
Fit Vector Autoregression model for benchmark comparison.
}
