% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annem_utils.R
\name{get_annem_config}
\alias{get_annem_config}
\title{Get ANNEM Configuration}
\usage{
get_annem_config()
}
\value{
List of configuration parameters
}
\description{
Return default configuration parameters for ANNEM simulations.
}
\examples{
config <- get_annem_config()
print(config)

}
