% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annem_utils.R
\name{set_annem_seed}
\alias{set_annem_seed}
\title{Set Random Seed for ANNEM}
\usage{
set_annem_seed(seed = 42)
}
\arguments{
\item{seed}{Integer random seed}
}
\description{
Set random seed for reproducible ANNEM simulations.
}
\examples{
set_annem_seed(42)

}
